match require optimizations/fixes
