+++ match/lib/match.rb
@@ -1,37 +1,12 @@
-require 'match/options'
-require 'match/runner'
-require 'match/nuke'
-require 'match/utils'
-require 'match/table_printer'
-require 'match/git_helper'
-require 'match/generator'
-require 'match/setup'
-require 'match/encrypt'
-require 'match/spaceship_ensure'
-require 'match/change_password'
-
-require 'fastlane_core'
-require 'terminal-table'
-require 'spaceship'
-
-module Match
-  Helper = FastlaneCore::Helper # you gotta love Ruby: Helper.* should use the Helper class contained in FastlaneCore
-  UI = FastlaneCore::UI
-  ROOT = Pathname.new(File.expand_path('../..', __FILE__))
-  DESCRIPTION = "Easily sync your certificates and profiles across your team using git"
-
-  def self.environments
-    return %w(appstore adhoc development enterprise)
-  end
-
-  def self.profile_type_sym(type)
-    return type.to_sym
-  end
-
-  def self.cert_type_sym(type)
-    return :enterprise if type == "enterprise"
-    return :development if type == "development"
-    return :distribution if ["adhoc", "appstore", "distribution"].include?(type)
-    raise "Unknown cert type: '#{type}'"
-  end
-end
+require_relative 'match/options'
+require_relative 'match/runner'
+require_relative 'match/nuke'
+require_relative 'match/utils'
+require_relative 'match/table_printer'
+require_relative 'match/git_helper'
+require_relative 'match/generator'
+require_relative 'match/setup'
+require_relative 'match/encrypt'
+require_relative 'match/spaceship_ensure'
+require_relative 'match/change_password'
+require_relative 'match/module'
