+++ match/lib/match/change_password.rb
@@ -1,3 +1,7 @@
+require_relative 'module'
+require_relative 'encrypt'
+require_relative 'git_helper'
+
 module Match
   # These functions should only be used while in (UI.) interactive mode
   class ChangePassword
