+++ match/lib/match/git_helper.rb
@@ -1,3 +1,8 @@
+require 'fastlane_core/command_executor'
+require_relative 'module'
+require_relative 'change_password'
+require_relative 'encrypt'
+
 module Match
   class GitHelper
     MATCH_VERSION_FILE_NAME = "match_version.txt"
