+++ match/lib/match/module.rb
@@ -1,0 +1,23 @@
+require 'fastlane_core/helper'
+
+module Match
+  Helper = FastlaneCore::Helper # you gotta love Ruby: Helper.* should use the Helper class contained in FastlaneCore
+  UI = FastlaneCore::UI
+  ROOT = Pathname.new(File.expand_path('../..', __FILE__))
+  DESCRIPTION = "Easily sync your certificates and profiles across your team using git"
+
+  def self.environments
+    return %w(appstore adhoc development enterprise)
+  end
+
+  def self.profile_type_sym(type)
+    return type.to_sym
+  end
+
+  def self.cert_type_sym(type)
+    return :enterprise if type == "enterprise"
+    return :development if type == "development"
+    return :distribution if ["adhoc", "appstore", "distribution"].include?(type)
+    raise "Unknown cert type: '#{type}'"
+  end
+end
