+++ match/lib/match/table_printer.rb
@@ -1,3 +1,9 @@
+require 'terminal-table'
+
+require 'fastlane_core/print_table'
+require_relative 'module'
+require_relative 'utils'
+
 module Match
   class TablePrinter
     # logs public key's  name, user, organisation, country, availability dates
